requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
bleach==6.3.0
click==8.3.1
Deprecated==1.3.1
fastapi==0.129.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.22
slowapi==0.1.9
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
webencodings==0.5.1
wrapt==2.1.1
